Refactor add() function
Extract the two main halves of add() into separate functions: one to
display the page for a certain lexeme (dropping the unused `total`
argument), one to handle a POST for it. We should be able to reuse both
for #11, with a different “driver” function to select the lexeme.

Also includes some other minor improvements: use flask.jsonify(), and
fix CSRF error detection (had no effect previously, now the form will be
repeated – with no special indication so far, but it’s something). Also,
submit_lexeme() no longer returns anything – instead of issuing a
redirect to the current URL (effectively asking the user to reload the
page), just fall through to responding with the page directly. (Unless
in “dev” mode, where the JSON is returned.)