Add route for lexemes by a user
This adds another route, /user/<user_name>, to pick a random lexeme
created by that user (typically your own, but you can also clean up
after someone else, or a bot account) which doesn’t have a sense yet.
The query to select the lexemes is adapted from T209034 [1].

If the query returns no results, either due to T209034 (incomplete
continuation) or because there really are no lexemes by that user
without senses, then the result is currently a server error. Perhaps
that can be improved later.

The route is not advertised anywhere yet, but I plan to add a link to
“your own lexemes” for logged-in users later.

Implements #11.

[1]: https://phabricator.wikimedia.org/T209034