Add link to current user’s lexemes to index page
Part of #11: when a user is logged in, show them a link to their own
lexemes without senses.

The phrasing and design is rubbish, I couldn’t think of anything better.
Feel free to replace it. (Perhaps include the number of such lexemes in
the description, and don’t show the link if it’s zero?)