Implement route for lexemes by a user (#13)
* Add hidden language input to lemma.html page

This means that POSTs to this page can be processed independently of the
route, which will be useful for #11.

* Set language on input

Some browsers enable spellchecking by default on single-line inputs, and
the user might also have enabled it manually, so let’s make sure the
browser checks spelling against the correct language.

* Refactor add() function

Extract the two main halves of add() into separate functions: one to
display the page for a certain lexeme (dropping the unused `total`
argument), one to handle a POST for it. We should be able to reuse both
for #11, with a different “driver” function to select the lexeme.

Also includes some other minor improvements: use flask.jsonify(), and
fix CSRF error detection (had no effect previously, now the form will be
repeated – with no special indication so far, but it’s something). Also,
submit_lexeme() no longer returns anything – instead of issuing a
redirect to the current URL (effectively asking the user to reload the
page), just fall through to responding with the page directly. (Unless
in “dev” mode, where the JSON is returned.)

* Add route for lexemes by a user

This adds another route, /user/<user_name>, to pick a random lexeme
created by that user (typically your own, but you can also clean up
after someone else, or a bot account) which doesn’t have a sense yet.
The query to select the lexemes is adapted from T209034 [1].

If the query returns no results, either due to T209034 (incomplete
continuation) or because there really are no lexemes by that user
without senses, then the result is currently a server error. Perhaps
that can be improved later.

The route is not advertised anywhere yet, but I plan to add a link to
“your own lexemes” for logged-in users later.

Implements #11.

[1]: https://phabricator.wikimedia.org/T209034

* Extract function to get and cache logged in user name

For the index function, we’ll need this information more than once, so
to avoid making two requests to the 'identify' API, cache the result in
the application context [1]. (The application context is destroyed at
the end of each request, so this should not result in stale information
across requests.)

[1]: http://flask.pocoo.org/docs/1.0/appcontext/

* Add link to current user’s lexemes to index page

Part of #11: when a user is logged in, show them a link to their own
lexemes without senses.

The phrasing and design is rubbish, I couldn’t think of anything better.
Feel free to replace it. (Perhaps include the number of such lexemes in
the description, and don’t show the link if it’s zero?)

* Redirect user after submitting a sense

If we don’t redirect the user after they’ve submitted a sense, and
instead fall through to directly responding with another form, then an
attempt to reload the page (typically intended to get another lexeme)
will instead be interpreted by the browser as an attempt to submit the
form again, which does not make sense.

* Fix “give me another lemma” button target

Make sure it always links to the current page (i. e. reloads it),
instead of the 'add' endpoint, which is not appropriate when using the
'user' endpoint.