Fix CSRF vulnerability (#15)
* Split off full_url() function from current_url()

For an upcoming commit, we’ll need a way to get the full version of a
URL that’s not the current one, so extract a full_url() function.

* Reject POSTs from the wrong referrer

On Wikimedia Toolforge, a CSRF token is not sufficient to fully protect
against cross-site request forgery (at least not if different tools are
considered different sites): since all tools run under the same domain
(same origin), another tool can fetch the page from this tool, including
the CSRF token, and then make valid requests.

To protect against this, also check the HTTP Referer header of any POST
requests made against the tool’s API, and reject requests coming from
anywhere but the current tool. (Since this tool has multiple pages where
a POST may legitimately come from – language, user, possibly more in the
future – we only check that the referrer base matches, and accept any
URL beneath the tool’s index URL.)

Note that we still hand out a CSRF token to GET requests with any
Referer header, which is necessary because we want sense creation to
work when people were linked directly to a particular page from
somewhere else. This means that an attacker can fetch the CSRF token and
then send it to an attacker-controlled server, from where they can make
requests with fake Referer headers; however, those requests will not
have the user’s session data. The user’s session is stored by Flask in a
cookie with the HttpOnly flag set, which should make it impossible for
any attacker to steal the session and combine it with the CSRF token.